add the triton implementation for benchamrk in the eagle utils
