+++ python/sgl_kernel/eagle_utils.py
@@ -4,6 +4,8 @@ import math
 from typing import List, Optional
 
 import torch
+import triton
+import triton.language as tl
 from sgl_kernel.speculative import TreeMaskMode
 from sgl_kernel.speculative import (
     build_tree_kernel_efficient as sgl_build_tree_kernel_efficient,
@@ -129,4 +131,211 @@ def build_tree_kernel_efficient(
         retrieve_next_token,
         retrieve_next_sibling,
         draft_tokens,
+    )
+
+
+@triton.jit
+def sgl_build_tree_kernel_efficient_triton(
+    parent_list_ptr,
+    selected_index_ptr,
+    verified_seq_len_ptr,
+    seq_len_prefix_sum_ptr,
+    tree_mask_ptr,
+    positions_ptr,
+    retrieve_index_ptr,
+    retrieve_next_token_ptr,
+    retrieve_next_sibling_ptr,
+    topk: tl.constexpr,
+    depth: tl.constexpr,
+    draft_token_num: tl.constexpr,
+    tree_mask_mode: tl.constexpr,
+    batch_size: tl.constexpr,
+    parent_list_stride: tl.constexpr,
+    selected_index_stride: tl.constexpr,
+):
+    """
+    Triton kernel for building EAGLE tree structure.
+    Each program handles one batch item (batch_idx).
+    """
+    batch_idx = tl.program_id(0)
+
+    # Calculate seq_tree_idx
+    seq_len = tl.load(verified_seq_len_ptr + batch_idx)
+    seq_len_prefix_sum = tl.load(seq_len_prefix_sum_ptr + batch_idx)
+
+    # Cast initial value to match the dtype of loaded tensors to avoid type inconsistency
+    seq_tree_idx = (
+        tl.cast(draft_token_num * draft_token_num * batch_idx, seq_len.dtype)
+        + seq_len_prefix_sum * draft_token_num
+    )
+
+    positions_offset = batch_idx * draft_token_num
+    tl.store(positions_ptr + positions_offset, seq_len)
+
+    retrieve_index_offset = batch_idx * draft_token_num
+
+    # Build retrieval index structure (reverse loop from draft_token_num-1 to 1)
+    for i in range(draft_token_num - 1, 0, -1):
+        current_token_idx = retrieve_index_offset + i
+        tl.store(
+            retrieve_index_ptr + batch_idx * draft_token_num + i,
+            current_token_idx,
+        )
+
+        parent_tb_idx = (
+            tl.load(selected_index_ptr + batch_idx * selected_index_stride + (i - 1))
+            // topk
+        )
+        parent_position = 0
+        found = 0
+
+        if parent_tb_idx == 0:
+            found = 1
+        else:
+            parent_token_idx = tl.load(
+                parent_list_ptr + batch_idx * parent_list_stride + parent_tb_idx
+            )
+
+            # Find parent position
+            for pp in range(draft_token_num - 1):
+                if found == 0:
+                    sel_idx = tl.load(
+                        selected_index_ptr + batch_idx * selected_index_stride + pp
+                    )
+                    if sel_idx == parent_token_idx:
+                        parent_position = pp + 1
+                        found = 1
+
+        if found == 1:
+            # Update next token links
+            next_tok_addr = (
+                retrieve_next_token_ptr + batch_idx * draft_token_num + parent_position
+            )
+            next_tok = tl.load(next_tok_addr)
+
+            if next_tok == -1:
+                tl.store(next_tok_addr, i)
+            else:
+                tl.store(next_tok_addr, i)
+                tl.store(
+                    retrieve_next_sibling_ptr + batch_idx * draft_token_num + i,
+                    next_tok,
+                )
+
+    tl.store(retrieve_index_ptr + batch_idx * draft_token_num, retrieve_index_offset)
+
+    # Process all draft token indices for tree mask
+    for draft_tokenx in range(draft_token_num):
+        if tree_mask_mode == 0:  # FULL_MASK
+            token_tree_idx = (
+                seq_tree_idx + (seq_len + draft_token_num) * draft_tokenx + seq_len + 1
+            )
+        else:
+            token_tree_idx = (
+                draft_token_num * draft_token_num * batch_idx
+                + draft_token_num * draft_tokenx
+                + 1
+            )
+
+        tl.store(tree_mask_ptr + token_tree_idx - 1, 1)
+        for i in range(draft_token_num - 1):
+            tl.store(tree_mask_ptr + token_tree_idx + i, 0)
+
+        if draft_tokenx > 0:
+            # Build tree path for draft_tokenx > 0
+            cur_position = draft_tokenx - 1
+            position = 0
+            should_continue = 1
+
+            for _ in range(depth):
+                if should_continue:
+                    position += 1
+                    tl.store(tree_mask_ptr + token_tree_idx + cur_position, 1)
+
+                    parent_tb_idx = (
+                        tl.load(
+                            selected_index_ptr
+                            + batch_idx * selected_index_stride
+                            + cur_position
+                        )
+                        // topk
+                    )
+                    if parent_tb_idx == 0:
+                        should_continue = 0
+                    else:
+                        parent_token_idx = tl.load(
+                            parent_list_ptr
+                            + batch_idx * parent_list_stride
+                            + parent_tb_idx
+                        )
+
+                        # Find cur_position for next iteration
+                        found = 0
+                        for cp in range(draft_token_num - 1):
+                            if found == 0:
+                                if (
+                                    tl.load(
+                                        selected_index_ptr
+                                        + batch_idx * selected_index_stride
+                                        + cp
+                                    )
+                                    == parent_token_idx
+                                ):
+                                    cur_position = cp
+                                    found = 1
+                        if found == 0:
+                            should_continue = 0
+
+            tl.store(
+                positions_ptr + batch_idx * draft_token_num + draft_tokenx,
+                position + seq_len,
+            )
+
+
+def sgl_build_tree_kernel_triton(
+    parent_list: torch.Tensor,
+    selected_index: torch.Tensor,
+    verified_seq_len: torch.Tensor,
+    tree_mask: torch.Tensor,
+    positions: torch.Tensor,
+    retrieve_index: torch.Tensor,
+    retrieve_next_token: torch.Tensor,
+    retrieve_next_sibling: torch.Tensor,
+    topk: int,
+    depth: int,
+    draft_token_num: int,
+    tree_mask_mode: TreeMaskMode = TreeMaskMode.FULL_MASK,
+):
+    """Triton-based implementation."""
+    # TODO: Add support for QLEN_ONLY_BITPACKING mode
+    if tree_mask_mode == TreeMaskMode.QLEN_ONLY_BITPACKING:
+        raise NotImplementedError(
+            "QLEN_ONLY_BITPACKING is not supported in Triton implementation"
+        )
+
+    batch_size = verified_seq_len.shape[0]
+    seq_len_prefix_sum = torch.cumsum(verified_seq_len, dim=0) - verified_seq_len
+
+    # Launch kernel with one program per batch item
+    grid = (batch_size,)
+
+    sgl_build_tree_kernel_efficient_triton[grid](
+        parent_list,
+        selected_index,
+        verified_seq_len,
+        seq_len_prefix_sum,
+        tree_mask,
+        positions,
+        retrieve_index,
+        retrieve_next_token,
+        retrieve_next_sibling,
+        topk=topk,
+        depth=depth,
+        draft_token_num=draft_token_num,
+        tree_mask_mode=int(tree_mask_mode),
+        batch_size=batch_size,
+        parent_list_stride=(
+            parent_list.stride(0) if parent_list.dim() > 1 else parent_list.shape[0]
+        ),
+        selected_index_stride=selected_index.stride(0),
     )
